requirements.txt
pulsectl-asyncio
python-rtmidi
dbus-next
pyyaml